# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: DamDam API
  version: 0.1.0
paths:
...
  /v1/me/account:
    delete:
      tags:
      - pilgrim-profile
      summary: Request Account Deletion
      operationId: request_account_deletion_v1_me_account_delete
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDeletionResponse'
      security:
      - HTTPBearer: []
  /v1/me/device-token:
    put:
      tags:
      - pilgrim-profile
      summary: Register Device Token
      operationId: register_device_token_v1_me_device_token_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceTokenUpsert'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceTokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /v1/me/family-contact:
    post:
      tags:
      - pilgrim-profile
      summary: Create Family Contact
      operationId: create_family_contact_v1_me_family_contact_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FamilyContactCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FamilyContactResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
    patch:
      tags:
      - pilgrim-profile
      summary: Update Family Contact
      operationId: update_family_contact_v1_me_family_contact_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FamilyContactUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FamilyContactResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /v1/me/emergency-contact:
    get:
      tags:
      - pilgrim-profile
      summary: Get Emergency Contact
      operationId: get_emergency_contact_v1_me_emergency_contact_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmergencyContactResponse'
      security:
      - HTTPBearer: []
...
  /v1/packages/purchase:
    post:
      tags:
      - payments
      summary: Purchase
      operationId: purchase_v1_packages_purchase_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /v1/packages/{package_id}/status:
    get:
      tags:
      - payments
      summary: Package Status
      operationId: package_status_v1_packages__package_id__status_get
      security:
      - HTTPBearer: []
      parameters:
      - name: package_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Package Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageStatusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/packages/{package_id}/geofence:
    get:
      tags:
      - payments
      summary: Package Geofence
      operationId: package_geofence_v1_packages__package_id__geofence_get
      security:
      - HTTPBearer: []
      parameters:
      - name: package_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Package Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageGeofenceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/webhooks/{processor}:
    post:
      tags:
      - payments
      summary: Webhook
      operationId: webhook_v1_webhooks__processor__post
      parameters:
      - name: processor
        in: path
        required: true
        schema:
          type: string
          title: Processor
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/checkins:
    post:
      tags:
      - check-ins
      summary: Create Checkin
      operationId: create_checkin_v1_checkins_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckInCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckInCreateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /v1/me/checkins:
    get:
      tags:
      - check-ins
      summary: History
      operationId: history_v1_me_checkins_get
      security:
      - HTTPBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 10
          minimum: 1
          default: 10
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckInHistoryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/webhooks/meta/whatsapp:
    get:
      tags:
      - check-ins
      summary: Verify Meta Webhook
      operationId: verify_meta_webhook_v1_webhooks_meta_whatsapp_get
      parameters:
      - name: hub.mode
        in: query
        required: true
        schema:
          type: string
          title: Hub.Mode
      - name: hub.verify_token
        in: query
        required: true
        schema:
          type: string
          title: Hub.Verify Token
      - name: hub.challenge
        in: query
        required: true
        schema:
          type: string
          title: Hub.Challenge
      responses:
        '200':
          description: Successful Response
          content:
            text/plain:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - check-ins
      summary: Meta Webhook
      operationId: meta_webhook_v1_webhooks_meta_whatsapp_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetaWebhookResponse'
...
components:
  schemas:
    AccountDeletionResponse:
      properties:
        status:
          type: string
          const: pending_deletion
          title: Status
          default: pending_deletion
        deletion_scheduled_for:
          type: string
          format: date-time
          title: Deletion Scheduled For
      type: object
      required:
      - deletion_scheduled_for
      title: AccountDeletionResponse
...
    CheckInCreate:
      properties:
        client_generated_id:
          type: string
          format: uuid
          title: Client Generated Id
        timestamp:
          type: string
          format: date-time
          title: Timestamp
        latitude:
          anyOf:
          - type: number
            maximum: 90.0
            minimum: -90.0
          - type: 'null'
          title: Latitude
        longitude:
          anyOf:
          - type: number
            maximum: 180.0
            minimum: -180.0
          - type: 'null'
          title: Longitude
      type: object
      required:
      - client_generated_id
      - timestamp
      title: CheckInCreate
    CheckInCreateResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        received_at:
          type: string
          format: date-time
          title: Received At
      type: object
      required:
      - id
      - received_at
      title: CheckInCreateResponse
    CheckInHistoryResponse:
      properties:
        checkins:
          items:
            $ref: '#/components/schemas/CheckInResponse'
          type: array
          title: Checkins
      type: object
      required:
      - checkins
      title: CheckInHistoryResponse
    CheckInResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        timestamp:
          type: string
          format: date-time
          title: Timestamp
        latitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Latitude
        longitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Longitude
      type: object
      required:
      - id
      - timestamp
      - latitude
      - longitude
      title: CheckInResponse
...
    DeviceTokenResponse:
      properties:
        registered:
          type: boolean
          title: Registered
          default: true
      type: object
      title: DeviceTokenResponse
    DeviceTokenUpsert:
      properties:
        fcm_token:
          type: string
          maxLength: 4096
          minLength: 20
          title: Fcm Token
        platform:
          $ref: '#/components/schemas/Platform'
      type: object
      required:
      - fcm_token
      - platform
      title: DeviceTokenUpsert
    EmergencyContactResponse:
      properties:
        hto_operator_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Hto Operator Name
        hto_operator_phone_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Hto Operator Phone Number
      type: object
      required:
      - hto_operator_name
      - hto_operator_phone_number
      title: EmergencyContactResponse
...
    FamilyContactCreate:
      properties:
        phone_number:
          type: string
          title: Phone Number
        name:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Name
        locale:
          $ref: '#/components/schemas/Locale'
          default: en
      type: object
      required:
      - phone_number
      title: FamilyContactCreate
    FamilyContactResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        phone_number:
          type: string
          title: Phone Number
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        notified_of_nomination:
          type: boolean
          title: Notified Of Nomination
        locale:
          $ref: '#/components/schemas/Locale'
      type: object
      required:
      - id
      - phone_number
      - name
      - notified_of_nomination
      - locale
      title: FamilyContactResponse
    FamilyContactUpdate:
      properties:
        phone_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone Number
        name:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Name
        locale:
          anyOf:
          - $ref: '#/components/schemas/Locale'
          - type: 'null'
      type: object
      title: FamilyContactUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Locale:
      type: string
      enum:
      - en
      - fr
      title: Locale
...
    MetaWebhookResponse:
      properties:
        processed:
          type: integer
          title: Processed
      type: object
      required:
      - processed
      title: MetaWebhookResponse
...
    PackageGeofenceResponse:
      properties:
        latitude:
          type: number
          title: Latitude
        longitude:
          type: number
          title: Longitude
        radius_meters:
          type: number
          title: Radius Meters
        request_id:
          type: string
          title: Request Id
      type: object
      required:
      - latitude
      - longitude
      - radius_meters
      - request_id
      title: PackageGeofenceResponse
    PackageStatusResponse:
      properties:
        status:
          type: string
          enum:
          - pending
          - active
          - expired
          title: Status
        data_gb_total:
          type: integer
          title: Data Gb Total
        data_gb_remaining:
          type: number
          title: Data Gb Remaining
        pstn_minutes_total:
          type: integer
          title: Pstn Minutes Total
        pstn_minutes_remaining:
          type: number
          title: Pstn Minutes Remaining
      type: object
      required:
      - status
      - data_gb_total
      - data_gb_remaining
      - pstn_minutes_total
      - pstn_minutes_remaining
      title: PackageStatusResponse
...
    Platform:
      type: string
      enum:
      - ios
      - android
      title: Platform
...
    PurchaseRequest:
      properties:
        pricing_tier_id:
          type: string
          format: uuid
          title: Pricing Tier Id
        group_size:
          anyOf:
          - type: integer
          - type: 'null'
          title: Group Size
      type: object
      required:
      - pricing_tier_id
      title: PurchaseRequest
    PurchaseResponse:
      properties:
        package_id:
          type: string
          format: uuid
          title: Package Id
        processor:
          type: string
          enum:
          - paystack
          - flutterwave
          title: Processor
        processor_reference:
          type: string
          title: Processor Reference
        checkout_url:
          type: string
          title: Checkout Url
      type: object
      required:
      - package_id
      - processor
      - processor_reference
      - checkout_url
      title: PurchaseResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    WebhookResponse:
      properties:
        processed:
          type: boolean
          title: Processed
      type: object
      required:
      - processed
      title: WebhookResponse
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer